Moves TemporalParse classes under scfg package.
